requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
python-dateutil==2.8.2
reportlab==4.0.4
